ml/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pillow==10.4.0
ultralytics==8.3.0
python-multipart==0.0.9
